Use 64 bit keys for TTable + replacement policy cleanups (#135)
There's also some changes to age and depth that doesn't affect the bench like the key size changes did. Those are done to make the code make more sense.

```
Advancement Test:
Time Control: 10s + 0.1s
Games: N=482 W=244 L=153 D=85
Elo: 66.4 +/- 28.6
Bench: 26355345

Alpha: 0.05
Beta: 0.05
Elo0: 0
Elo1: 10
H1 was accepted
```

+++ src/ttable.cpp
@@ -34,7 +34,7 @@ int TTable::hashFull() {
 
 bool TTable::entryExists(uint64_t key) const {
     int index = this->getIndex(key);
-    return static_cast<uint16_t>(key) == this->table[index].key;
+    return key == this->table[index].key;
 }
 
 Entry TTable::getEntry(int index) const {
